Migrated model
